Move class TaskQueue into new header
